# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Joy VoyagerApi module adds REST Api end points to Voyager with Passport and Swagger support https://github.com/rxcod9/joy-voyager-api.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://finishuseruz.demo",
            "description": "Primary API Server"
        },
        {
            "url": "http://my-default-host.com",
            "description": "Another API Server"
        }
    ],
    "paths": {
...
        "/api/ref": {
            "get": {
                "tags": [
                    "Refill"
                ],
                "summary": "Get list of Refill",
                "operationId": "dedba959969048c9dffcbf95b3f3db7b",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "500": {
                        "description": "Ajax error"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/prepare": {
            "post": {
                "tags": [
                    "Refill"
                ],
                "summary": "Create new Prepare",
                "operationId": "be047d6d041f9b738528f406ac67ce1d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "click_trans_id": {
                                                "type": "integer"
                                            },
                                            "service_id": {
                                                "type": "integer"
                                            },
                                            "click_paydoc_id": {
                                                "type": "integer"
                                            },
                                            "merchant_trans_id": {
                                                "type": "integer"
                                            },
                                            "amount": {
                                                "type": "integer"
                                            },
                                            "action": {
                                                "type": "integer"
                                            },
                                            "error": {
                                                "type": "integer"
                                            },
                                            "error_note": {
                                                "type": "string"
                                            },
                                            "sign_time": {
                                                "type": "string"
                                            },
                                            "sign_string": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "click_trans_id": 1613910471,
                                    "service_id": 19839,
                                    "click_paydoc_id": 1747426201,
                                    "merchant_trans_id": 201,
                                    "amount": 6000,
                                    "action": 0,
                                    "error": 0,
                                    "error_note": "Success",
                                    "sign_time": "2022-02-14 20:56:40",
                                    "sign_string": "0fc2457d6db176d12f96b8af2aa25679"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "click_trans_id": {
                                            "type": "integer",
                                            "example": 1613910471
                                        },
                                        "service_id": {
                                            "type": "integer",
                                            "example": 19839
                                        },
                                        "click_paydoc_id": {
                                            "type": "integer",
                                            "example": 1747426201
                                        },
                                        "merchant_trans_id": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": 6000
                                        },
                                        "action": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "error": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "error_note": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "sign_time": {
                                            "type": "string",
                                            "example": "2022-02-14 20:56:40"
                                        },
                                        "sign_string": {
                                            "type": "string",
                                            "example": "0fc2457d6db176d12f96b8af2aa25679"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/complete": {
            "post": {
                "tags": [
                    "Refill"
                ],
                "summary": "Create new Complete",
                "operationId": "76b5301146da4c9c5a3a45fb55809b5f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "click_trans_id": {
                                                "type": "integer"
                                            },
                                            "service_id": {
                                                "type": "integer"
                                            },
                                            "click_paydoc_id": {
                                                "type": "integer"
                                            },
                                            "merchant_trans_id": {
                                                "type": "integer"
                                            },
                                            "merchant_prepare_id": {
                                                "type": "integer"
                                            },
                                            "amount": {
                                                "type": "integer"
                                            },
                                            "action": {
                                                "type": "integer"
                                            },
                                            "error": {
                                                "type": "integer"
                                            },
                                            "error_note": {
                                                "type": "string"
                                            },
                                            "sign_time": {
                                                "type": "string"
                                            },
                                            "sign_string": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "click_trans_id": 1613910471,
                                    "service_id": 19839,
                                    "click_paydoc_id": 1747426201,
                                    "merchant_trans_id": 201,
                                    "merchant_prepare_id": 33,
                                    "amount": 6000,
                                    "action": 1,
                                    "error": 0,
                                    "error_note": "Success",
                                    "sign_time": "2022-02-14 20:56:40",
                                    "sign_string": "0fc2457d6db176d12f96b8af2aa25679"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "click_trans_id": {
                                            "type": "integer",
                                            "example": 1613910471
                                        },
                                        "service_id": {
                                            "type": "integer",
                                            "example": 19839
                                        },
                                        "click_paydoc_id": {
                                            "type": "integer",
                                            "example": 1747426201
                                        },
                                        "merchant_trans_id": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "merchant_prepare_id": {
                                            "type": "integer",
                                            "example": 33
                                        },
                                        "amount": {
                                            "type": "integer",
                                            "example": 6000
                                        },
                                        "action": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "error": {
                                            "type": "integer",
                                            "example": 0
                                        },
                                        "error_note": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "sign_time": {
                                            "type": "string",
                                            "example": "2022-02-14 20:56:40"
                                        },
                                        "sign_string": {
                                            "type": "string",
                                            "example": "0fc2457d6db176d12f96b8af2aa25679"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "example": "2021-12-11T09:25:53.000000Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "invalid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "msg": {
                                            "type": "string",
                                            "example": "fail"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "token": {
                "type": "apiKey",
                "description": "Bearer token",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}